+++ example/lib/custom_text_overflow_demo.dart
@@ -14,10 +14,10 @@ class CustomTextOverflowDemo extends StatelessWidget {
         body: Container(
           padding: EdgeInsets.all(20.0),
           child: ExtendedText(
-            "relate to \$issue 26748\$ .[love]]Extended text help you to build rich text quickly. any special text you will have with extended text. "
+            "relate to \$issue 26748\$ .[love]Extended text help you to build rich text quickly. any special text you will have with extended text. "
             "It's my pleasure to invite you to join \$FlutterCandies\$ if you want to improve flutter .[love]"
             "if you meet any problem, please let me konw @zmtzawqlp .[sun_glasses] "
-            "relate to \$issue 26748\$ .[love]]Extended text help you to build rich text quickly. any special text you will have with extended text. "
+            "relate to \$issue 26748\$ .[love]Extended text help you to build rich text quickly. any special text you will have with extended text. "
             "It's my pleasure to invite you to join \$FlutterCandies\$ if you want to improve flutter .[love]"
             "if you meet any problem, please let me konw @zmtzawqlp .[sun_glasses] ",
             onSpecialTextTap: (dynamic parameter) {